requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==3.2.19
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-writable-nested==0.7.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.11.1
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
PyJWT==2.7.0
pytest==7.3.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.0.3
zipp==3.15.0
